_includes/resources/azure_functions/durable_testing_python/requirements.txt
azure-functions~=1.8.0
azure-functions-durable~=1.1.3
pytest~=6.2.5
